remove r2d2 shenanigans
